requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
beautifulsoup4==4.12.2
aiohttp==3.9.0
fake-useragent==1.4.0
python-dotenv==1.0.0
